Fixed problem with guzzle's new version
Changed the "body" request to "form_params". This is fixed that error:

Passing in the "body" request option as an array to send a POST request
has been deprecated. Please use the "form_params" request option to send
a application/x-www-form-urlencoded request, or a the "multipart"
request option to send a multipart/form-data request.